fix/rename folder from execption to exception
