+++ src/main/java/com/delivery/auth/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.delivery.auth.service;
 
 import com.delivery.auth.dto.SignUpRequestDto;
-import com.delivery.execption.BusinessException;
-import com.delivery.execption.ErrorCode;
+import com.delivery.exception.BusinessException;
+import com.delivery.exception.ErrorCode;
 import com.delivery.user.entity.User;
 import com.delivery.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
