+++ src/main/java/com/delivery/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.delivery.execption;
+package com.delivery.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
